ENH: handle events triggered by the GUI elements for the connection setup
Added function setCollaborationNode to handle changes in the collaboration node selector.
Added function onConnectButtonClicked to active the connector when Connect button is clicked.
Added function updateConnectorNode to update the connector properties whenever a property is modified from the GUI.
Added function updateWidgetFromMRML to update the GUI values when the selected collaboration node is modified.