ENH: modified the attribute filter to be the ID of the selected collaboration node; updated the attribute filter every time the collaboration node changes
The variable SELECTED_COLLAB_NODE has been added to store the name of the selected collaboration node. If no node is selected, it has the value "None". If a node is selected, it takes the value of the collaboration node's ID.
Whenever a collaboration node is selected, the current attribute filter of the tree view is removed, and a new attribute filter is created with the name of SELECTED_COLLAB_NODE.