ENH: Handle sending nodes on connect and after connection
Add attribute ("OpenIGTLinkIF.pushOnConnect", "true") to nodes selected for synchronization. When the connection is established, PushNodes() is called to send them.
New nodes added for synchronization while the connection is active are also sent.
The button Send has been added to be able to send all nodes again. This is useful when clients connect after the server, so they receive nodes that are not being modified.