ENH: Override connectorNode functions to send DisplayNodes through TextNodes
Functions CreateNewMRMLNodeForDevice and ProcessIncomingDeviceModifiedEvent are overriden to convert text nodes containing display node attributes (as an xml text) to display nodes, and assign them to the corresponding model node.

In synchronizeSelectedNodes function from the widget, if a model node is selected for synchronization, a text node is created and sent containing the display node properties as an XML. If it is no longer synchronized, the text node is removed.