ENH: send transforms information in a text node
When a transform is selected for synchronization, a text node is generated containing the names of the nodes observing that transform.
The client receives the transform and text nodes and makes the nodes observe the transform,
The transforms are correctly applied independently of the order in which nodes are received.